CB#7 - Refactoring - Fix structure in projects Database and Domain
