+++ src/crypto-bank.Database/DatabaseProjectServiceCollectionExtensions.cs
@@ -3,7 +3,7 @@ using Microsoft.Extensions.DependencyInjection;
 
 namespace crypto_bank.Database;
 
-public static class ServiceCollectionExtensions
+public static class DatabaseProjectServiceCollectionExtensions
 {
     public static IServiceCollection AddDatabaseProject(this IServiceCollection services)
     {
