+++ src/crypto-bank.Domain/Models/User.cs
@@ -1,6 +1,6 @@
 using crypto_bank.Domain.Authorization;
 
-namespace crypto_bank.Domain.Features.Users.Models;
+namespace crypto_bank.Domain.Models;
 
 public record User
 {
