+++ src/crypto-bank.WebAPI/Features/Users/Requests/UpdateRoles.RequestHandler.cs
@@ -1,5 +1,5 @@
 using crypto_bank.Database;
-using crypto_bank.Domain.Features.Users.Models;
+using crypto_bank.Domain.Models;
 using JetBrains.Annotations;
 using MediatR;
